requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.12
django-admin-autocomplete-filter==0.7.1
django-admin-rangefilter==0.8.3
django-baton==2.2.3
django-filer==2.1.2
django-import-export==2.7.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-polymorphic==3.0.0
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
lxml==4.8.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
PyYAML==6.0
reportlab==3.6.6
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.0
tinycss2==1.1.1
tzdata==2021.5
Unidecode==1.1.2
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
